Update mcwrap.py to work with Prism Launcher.
The original mcwrap.py works perfectly on MultiMC, but on multiple computers and with multiple reinstalls, would not play nicely with Prism Launcher. This modification simply checks if the directory exists on launch of the Prism Launcher instance, and if it doesn't, it creates the required directory, allowing the script to work properly.